warn instead or raise an error when doing an action before initialization
